requirements (1).txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.28.0
scikit-learn>=1.3.0
tensorflow-cpu>=2.13.0
folium>=0.14.0
streamlit-folium>=0.15.0
plotly>=5.18.0
